requirements.txt
flask
openai
yt-dlp
gspread
google-auth
isodate
